There is already a rubocop for this
