+++ test/testdata/packager/simple_test_import/main_lib/lib.rb
@@ -4,13 +4,14 @@
 class Project::MainLib::Lib
   Project::Util::MyUtil.new
 
-  # Normal code is not allowed to access names from `test_import`
   Project::TestOnly::SomeHelper.new
-# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: Used `test_import` constant `Project::TestOnly::SomeHelper` in non-test file
+# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: `Project::TestOnly::SomeHelper` resolves but its package is not imported
 
   Test::Project::Util::UtilHelper
-# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: `Test::Project::Util::UtilHelper` is defined in a test namespace
+# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: `Test::Project::Util::UtilHelper` resolves but its package is not imported
+# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: Package `Project::MainLib` may not reference `test!` packages
 
   Test::Project::Util::Unexported
-# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: `Test::Project::Util::Unexported` is defined in a test namespace
+# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: `Test::Project::Util::Unexported` resolves but is not exported from `Test::Project::Util` and `Test::Project::Util` is not imported
+# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: Package `Project::MainLib` may not reference `test!` packages
 end
