+++ test/testdata/packager/simple_test_import/main_lib/test/__package.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+# typed: strict
+
+class Test::Project::MainLib < PackageSpec
+  test!
+
+  import Project::MainLib, uses_internals: true
+  import Project::Util
+  import Project::TestOnly
+  import Test::Project::Util
+end
