+++ test/testdata/packager/simple_test_import/util/__package.rb
@@ -4,5 +4,4 @@
 
 class Project::Util < PackageSpec
   export Project::Util::MyUtil
-  export Test::Project::Util::UtilHelper
 end
