+++ test/testdata/packager/simple_test_import/util/test/__package.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+# typed: strict
+
+class Test::Project::Util < PackageSpec
+  test!
+
+  import Project::Util
+
+  export Test::Project::Util::UtilHelper
+end
