+++ test/testdata/packager/test-packages-bad-test/__package.rb
@@ -3,7 +3,6 @@
 
 class Root < PackageSpec
   test!
-# ^^^^^ error: `test!` is only valid for packages with `/test/` in their path
 
   export Root::A
 end
