+++ test/testdata/packager/test-packages-bad-test/b/__package.rb
@@ -3,5 +3,4 @@
 
 class Test::Root::B < PackageSpec
   test!
-# ^^^^^ error: `test!` is only valid for packages with `/test/` in their path
 end
