+++ test/testdata/packager/test-packages-min-levels/__package.rb
@@ -3,7 +3,7 @@
 
 class Root < PackageSpec
   sorbet min_typed_level: 'true', tests_min_typed_level: 'true'
-  #                               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: Invalid expression in package
+  #                               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: Unrecognized keyword argument `tests_min_typed_level`
 
   export Root::A
 end
