+++ test/testdata/packager/test-packages-min-levels/test/__package.rb
@@ -4,7 +4,7 @@ class Test::Root < PackageSpec
   test!
 
   sorbet tests_min_typed_level: 'strict', min_typed_level: 'strict'
-  #      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: Invalid expression in package
+  #      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ error: Unrecognized keyword argument `tests_min_typed_level`
 
 
   import Root, uses_internals: true
